🚚rename: 모든 폴더 명 뒤에 F 추가
