+++ frontend/src/routes/rewardPageF/receiptF/Receipt.jsx
