+++ frontend/src/routes/rewardPageF/tutorialF/receiptF/ReceiptTutorial.jsx
