+++ frontend/src/routes/rewardPageF/tutorialF/receiptF/ReceiptTutorial.module.css
