+++ frontend/src/routes/stockPageF/StockLayout.jsx
