+++ frontend/src/routes/stockPageF/StockPage.jsx
